move game and chat gateway function into wsocket gateway
