+++ backend/src/chat/chat.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ChatGateway } from './chat.gateway';
+import { WSocketGateway } from '../wsocket/wsocket.gateway';
 import { ChatService } from './chat.service';
 import { ChatController } from './chat.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -10,7 +10,7 @@ import { ChatDAO } from './chat.dao';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, Messages, Channels])],
-  providers: [ChatGateway, ChatService, UsersService, ChatDAO],
+  providers: [WSocketGateway, ChatService, UsersService, ChatDAO],
   controllers: [ChatController],
 })
 export class ChatModule {}
