+++ backend/src/chat/chat.service.ts
@@ -1,7 +1,13 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { User } from 'src/users/user.entity';
 import { UsersService } from 'src/users/users.service';
+import { IMessage } from './properties';
 import { ChatDAO } from './chat.dao';
+import { manageUsers, gameInvite, gameAccept } from './chat.gameinvite';
+import { Socket, Server } from 'socket.io';
+
+const rooms: string[] = [];
+const messages: Map<string, string[]> = new Map<string, string[]>();
 
 @Injectable()
 export class ChatService {
@@ -39,6 +45,29 @@ export class ChatService {
     const user = await this.userService.findOneByName(userId);
     if (user) {
       return await this.userService.clearActiveChats(user);
+    }
+  }
+
+  joinRoom(data: IMessage, client: Socket) {
+    const mess = `${data.user.name} joined ${data.room}`;
+    manageUsers(data, client);
+    client.join(data.room);
+    if (!messages.has(data.room)) {
+      messages.set(data.room, []);
+      rooms.push(data.room);
+    }
+    client.to(data.room).emit('message', mess);
+    messages.get(data.room).push(mess);
+    return messages.get(data.room);
+
+  }
+
+  handleMessage( data: IMessage, client: Socket, server: Server) {
+    manageUsers(data, client);
+    const mess = `${data.user.name}: ${data.input}`;
+    if (!gameInvite(data, server) && !gameAccept(data, server)) {
+      server.to(data.room).emit('message', mess);
+      messages.get(data.room).push(mess);
     }
   }
 }
