+++ backend/src/games/games.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { GamesGateway } from './games.gateway';
 import { GamesService } from './games.service';
 import { Game } from './game.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { WSocketGateway } from '../wsocket/wsocket.gateway';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Game])],
-  providers: [GamesService, GamesGateway],
+  providers: [GamesService, WSocketGateway],
 })
 export class GamesModule {}
